pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "apptest"
version = "0.1.0"
description = "AI-powered test generation from PR diffs"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
apptest = "apptest.cli:main"

[tool.setuptools.packages.find]
include = ["apptest*"]

[tool.pytest.ini_options]
testpaths = ["apptest/tests"]
